rename: Log 관련 클래스 config.log 패키지로 이동 (#353)
