+++ module-presentation/src/main/java/com/depromeet/friend/api/FollowController.java
@@ -1,6 +1,6 @@
 package com.depromeet.friend.api;
 
-import com.depromeet.config.Logging;
+import com.depromeet.config.log.Logging;
 import com.depromeet.dto.response.ApiResponse;
 import com.depromeet.friend.dto.request.FollowRequest;
 import com.depromeet.friend.dto.response.*;
