+++ module-presentation/src/main/java/com/depromeet/pool/api/PoolController.java
@@ -1,6 +1,6 @@
 package com.depromeet.pool.api;
 
-import com.depromeet.config.Logging;
+import com.depromeet.config.log.Logging;
 import com.depromeet.dto.response.ApiResponse;
 import com.depromeet.member.annotation.LoginMember;
 import com.depromeet.pool.dto.request.FavoritePoolCreateRequest;
