+++ module-presentation/src/main/java/com/depromeet/reaction/api/ReactionController.java
@@ -1,6 +1,6 @@
 package com.depromeet.reaction.api;
 
-import com.depromeet.config.Logging;
+import com.depromeet.config.log.Logging;
 import com.depromeet.dto.response.ApiResponse;
 import com.depromeet.member.annotation.LoginMember;
 import com.depromeet.reaction.dto.request.ReactionCreateRequest;
